refactoring; check() doesn't complain about compilation
